[REF] product_variant_sale_price: split product models and rename test file
- Moved product.template methods to dedicated product_template.py
- Renamed test_product_product.py -> test_product.py since both models are involved in the tests

[UPD] Update product_variant_sale_price.pot

[BOT] post-merge updates
