+++ product_variant_sale_price/README.rst
@@ -1,3 +1,7 @@
+.. image:: https://odoo-community.org/readme-banner-image
+   :target: https://odoo-community.org/get-involved?utm_source=readme
+   :alt: Odoo Community Association
+
 ==========================
 Product Variant Sale Price
 ==========================
@@ -7,13 +11,13 @@ Product Variant Sale Price
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:47713ad2bb39e1e72427b7027485630d976380dad3b090db3b737712a26e90fa
+   !! source digest: sha256:fa91a9986c9ca50e710e86a15bf5c3fccc99a2f73e51d08cb44a16a45d874e9e
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
     :target: https://odoo-community.org/page/development-status
     :alt: Beta
-.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+.. |badge2| image:: https://img.shields.io/badge/license-AGPL--3-blue.png
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fproduct--variant-lightgray.png?logo=github
@@ -62,16 +66,16 @@ Authors
 Contributors
 ------------
 
--  Carlos Dauden <carlos.dauden@tecnativa.com>
--  Sergio Teruel <sergio.teruel@tecnativa.com>
--  Alex Comba <alex.comba@agilebg.com>
--  Fabien Bourgeois <fabien@yaltik.com>
--  Vicent Cubells <info@obertix.net>
--  RabbitJon-S73 <roger@studio73.es>
--  Emanuel Cino <ecino@compassion.ch>
--  Pedroguirao <pguirao@puntsistemes.es>
--  Gabriel Grinspan <gabriel@rooteam.net>
--  Nils Coenen <nils.coenen@nico-solutions.de>
+- Carlos Dauden <carlos.dauden@tecnativa.com>
+- Sergio Teruel <sergio.teruel@tecnativa.com>
+- Alex Comba <alex.comba@agilebg.com>
+- Fabien Bourgeois <fabien@yaltik.com>
+- Vicent Cubells <info@obertix.net>
+- RabbitJon-S73 <roger@studio73.es>
+- Emanuel Cino <ecino@compassion.ch>
+- Pedroguirao <pguirao@puntsistemes.es>
+- Gabriel Grinspan <gabriel@rooteam.net>
+- Nils Coenen <nils.coenen@nico-solutions.de>
 
 Maintainers
 -----------
