+++ product_variant_sale_price/models/__init__.py
@@ -2,3 +2,4 @@
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
 from . import product_product
+from . import product_template
