+++ product_variant_sale_price/tests/__init__.py
@@ -1,4 +1,4 @@
 # Copyright 2016 Sergio Teruel <sergio.teruel@tecnativa.com>
 # License AGPL-3 - See http://www.gnu.org/licenses/agpl-3.0.html
 
-from . import test_product_product
+from . import test_product
